pyproject.toml
[project]
name = "agentbeats-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.22",
    "httpx>=0.28.1",
    "starlette>=0.52.1",
    "uvicorn>=0.40.0",
]
